Adds main create_short_url service function
This is the main function of the service, it calls the helper functions to generate a slug, instantiates a new Link object from the slug and original long_url, and adds those to the database before returning the new short url.

It does a lot, and might be a candidate for refactoring in the future. However it is quite short already.