Adds tests for `get_long_url` service
This service is much smaller, all it does at the moment is get the matching long_url from the database using the slug as a key.

It raises a custom error if there is no matching slug.

The error is based on the advice given by ruff: https://docs.astral.sh/ruff/rules/raise-vanilla-args/