FIX: broken validation of URL length
It turns out Pydantic doesn't automatically check for length, or raise errors.

You have to write your own validation functions yourself.

This commit reads MIN and MAX lengths from the `.env` and applies them in a field_validator to pydantic. It raises errors if the length is too short or too long.